+++ src/Model/Creature/Imp.php
@@ -44,7 +44,7 @@ class Imp extends AbstractCreature
         } else {
             $this->name = self::COMMON_NAME.' - '.$this->getRawName();
         }
-        $this->health = ceil($this->stats->getEndurance() * ceil($scale / 2) * ($this->creatureClass->getValue() / 100));
+        $this->health = (int) ceil($this->stats->getEndurance() * ceil($scale / 2) * ($this->creatureClass->getValue() / 100));
         $this->experience = $this->createRandomNumberInRangeWithScale(5, 7, $scale);
     }
